not all functions work did a lotta stuff

+++ strlen.c
@@ -1,5 +1,5 @@
 size_t strlen(char* string) {
-	int i = 0;
+	uint16_t i = 0;
 	for(i = 0; string[i] != '\0'; i++) {}
 	return i;
 }
